Use Symbol in place of ImportedSymbol
This refactors the sniff to store symbol imports as instances of
`Symbol` rather than `ImportedSymbol`. `Symbol` is much more intelligent
and can keep track of the different parts of a symbol, rather than only
the imported alias. It can also identify full namespace paths even for
grouped imports.

More importantly, this marks each imported symbol as "used" directly as
soon as it is imported if it matches one of our ignored patterns. This
prevents issues where the pattern match would fail for symbols in an
import statement since `markSymbolUsed()` first tries to retrieve the
appropriate symbol from the records based on its top-level namespace.

That technique works fine for regular symbol use (eg: when looking to
see if `Foo\bar()` is imported, we need to look for a statement like
`use Lib\Classes\Foo`) but will fail for the symbol in the import
itself. In the latter case, if we mark it as "used" before it's stored
in the records, then we don't have to look it up at all.