feat(detection) Detect symbol of kind `abc\{def\ghi, klm\nop}`.
This patch updates `getImportedSymbolsFromGroupStatement` to
understand namespaced symbols (`x\y`) inside a group (`{…}`), like
`abc\{def\ghi, klm\nop}`.

This patch also adds, for the sake of completeness, an alias property
on `Symbol`, so that not only the last strings of a namespaced name is
considered as an alias, but the real alias string (`as x`, here `x`)
can be used as an alias if declared.