Add test case for automatic string conversion
This may be confusing because if PHP encounters an undefined symbol, it
will issue a warning and try converting it to a string. Because of this
fallback, using unquoted strings may seem like valid PHP, but it's not
doing what you might think it's doing; it's actually still using a
string. In a future version of php, this fallback may be removed, so
it's recommended to avoid it (see the warning here:
https://www.php.net/manual/en/migration72.deprecated.php).