Composer: PHPCS is a production dependency
PHP_CodeSniffer is a dependency for this external standard to be able to run, so should be in `require`, not `require-dev`.

This also makes sure that people won't be able to try and run the standard with incompatible PHPCS versions.

Includes making the version requirement more flexible (not fixed to one version `3.3.0` only).